# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Flight Radar - Laravel",
        "description": "This is a Rest API built with Laravel for tracking personal flight hidstory",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/flights": {
            "get": {
                "tags": [
                    "Flight"
                ],
                "summary": "Get a list of flights created by the user",
                "operationId": "4baffc280d4c3e239575b8c57e979397",
                "parameters": [
                    {
                        "name": "airline",
                        "in": "query",
                        "description": "Airline in ICAO Code",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "airport",
                        "in": "query",
                        "description": "Airport in IATA Code",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Token is invalid"
                    },
                    "404": {
                        "description": "Unable to retrieve user from database"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Flight"
                ],
                "summary": "Create a new flight",
                "operationId": "c34d49c92e19085907c2b95d755a93f6",
                "requestBody": {
                    "description": "Flight payload format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "departure_date": {
                                        "description": "Departure Date in ISO 8601 Format",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "flight_number": {
                                        "description": "Flight number",
                                        "type": "string"
                                    },
                                    "departure_airport": {
                                        "description": "Departure airport in IATA Code",
                                        "type": "string"
                                    },
                                    "arrival_airport": {
                                        "description": "Departure airport in IATA Code",
                                        "type": "string"
                                    },
                                    "distance": {
                                        "description": "Flight distance in miles",
                                        "type": "number"
                                    },
                                    "airline": {
                                        "description": "Airline in ICAO Code",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully created flight"
                    },
                    "401": {
                        "description": "Token is invalid"
                    },
                    "404": {
                        "description": "Unable to retrieve flight"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/flight-stats": {
            "get": {
                "tags": [
                    "Flight"
                ],
                "summary": "Get an abundance of flight stats from user's flight history",
                "operationId": "0eb3b95a664cd8792dfaa9b2cd796916",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Token is invalid"
                    },
                    "404": {
                        "description": "Unable to retrieve user from database"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/flights/{id}": {
            "get": {
                "tags": [
                    "Flight"
                ],
                "summary": "Get a specific flight",
                "operationId": "68077cee376d99dd1f625e4b5c0236b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Flight ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Token is invalid"
                    },
                    "404": {
                        "description": "Unable to retrieve flight"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Flight"
                ],
                "summary": "Update an existing flight",
                "operationId": "29d9b985eff7a7d40f56bd4a377854c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Flight ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Flight payload format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "departure_date": {
                                        "description": "Departure Date in ISO 8601 Format",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "flight_number": {
                                        "description": "Flight number",
                                        "type": "string"
                                    },
                                    "departure_airport": {
                                        "description": "Departure airport in IATA Code",
                                        "type": "string"
                                    },
                                    "arrival_airport": {
                                        "description": "Departure airport in IATA Code",
                                        "type": "string"
                                    },
                                    "distance": {
                                        "description": "Flight distance in miles",
                                        "type": "number"
                                    },
                                    "airline": {
                                        "description": "Airline in ICAO Code",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful updated flight"
                    },
                    "401": {
                        "description": "Token is invalid"
                    },
                    "404": {
                        "description": "Unable to retrieve flight"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Flight"
                ],
                "summary": "Delete a specific flight from user's flight list",
                "operationId": "59fd658ea7800ce51bbc12bb83c73c79",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Flight ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully deleted flight"
                    },
                    "401": {
                        "description": "Token is invalid"
                    },
                    "404": {
                        "description": "Unable to retrieve flight"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}